+++ packages/contracts/test/plugins/governance/multisig/multisig.ts
@@ -207,6 +207,14 @@ describe('Multisig', function () {
         .to.emit(multisig, MULTISIG_EVENTS.MULTISIG_SETTINGS_UPDATED)
         .withArgs(multisigSettings.onlyListed, multisigSettings.minApprovals);
     });
+
+    it('should revert if members list is longer than uint16 max', async () => {
+      const megaMember = signers[1];
+      const members: string[] = new Array(65537).fill(megaMember.address);
+      await expect(multisig.initialize(dao.address, members, multisigSettings))
+        .to.revertedWithCustomError(multisig, 'AddresslistLengthOutOfBounds')
+        .withArgs(65535, members.length);
+    });
   });
 
   describe('plugin interface: ', async () => {