+++ packages/contracts/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Created the `IMultisig` interface.
+- Adds check to `Multisig.initialize()` that members length isn't bigger than max uint16.
 
 ### Changed
 
@@ -36,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use OZ's upgradeable contracts for `PluginCloneable`.
 - Renamed `getDAO()` to `dao()` and changed the `dao` state variable mutability to private.
 - Disallows creating a new proposal in the Multisig plugin in the same block whereas the settings have been changed.
+- Updates `MultisigSetup` to be used as Release 1 Build 2.
 
 ### Removed